+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # egui_dock changelog
 
+## Unreleased
+
+### Added
+- `Style` now includes an option to change the tab's height - `tab_bar_height`. ([#62](https://github.com/Adanos020/egui_dock/pull/62))
+
 ## 0.3.0 - 2022-12-10
 
 ### Added